Refactor HttpClient creation into a builder and simplify NetworkModule
- Create `HttpClientBuilder` to centralize Ktor `HttpClient` configuration, including negotiation, timeouts, authentication, and error validation.
- Refactor `NetworkModule` to use `HttpClientBuilder` and remove redundant host/profile qualifiers.
- Clean up unused interceptor qualifiers in `QualifierNetworking.kt`.
